Only throw on login mutations
