+++ client/src/index.tsx
@@ -5,36 +5,27 @@ import AppAuthenticated from './AppAuthenticated'
 import AppUnauthenticated from './AppUnauthenticated'
 import AppWelcome from './AppWelcome'
 import 'semantic-ui-css/semantic.min.css'
-import { ReactQueryConfigProvider } from 'react-query'
 import { ReactQueryDevtools } from 'react-query-devtools'
 
-const queryConfig = {
-    mutations: {
-        throwOnError: true,
-    },
-}
-
 const routing = (
-    <ReactQueryConfigProvider config={queryConfig}>
-        <BrowserRouter>
-            <ReactQueryDevtools initialIsOpen={false} />
-            <Switch>
-                <Route exact path="/" component={AppUnauthenticated} />
-                <Route path="/signup" component={AppUnauthenticated} />
-                <Route path="/login" component={AppUnauthenticated} />
+    <BrowserRouter>
+        <ReactQueryDevtools initialIsOpen={false} />
+        <Switch>
+            <Route exact path="/" component={AppUnauthenticated} />
+            <Route path="/signup" component={AppUnauthenticated} />
+            <Route path="/login" component={AppUnauthenticated} />
 
-                <Route path="/home" component={AppAuthenticated} />
-                <Route path="/user/:username" component={AppAuthenticated} />
-                <Route path="/post/:postid" component={AppAuthenticated} />
-                <Route path="/followers" component={AppAuthenticated} />
-                <Route path="/following" component={AppAuthenticated} />
-                <Route path="/requests" component={AppAuthenticated} />
-                <Route path="/new" component={AppAuthenticated} />
+            <Route path="/home" component={AppAuthenticated} />
+            <Route path="/user/:username" component={AppAuthenticated} />
+            <Route path="/post/:postid" component={AppAuthenticated} />
+            <Route path="/followers" component={AppAuthenticated} />
+            <Route path="/following" component={AppAuthenticated} />
+            <Route path="/requests" component={AppAuthenticated} />
+            <Route path="/new" component={AppAuthenticated} />
 
-                <Route path="/welcome" component={AppWelcome} />
-            </Switch>
-        </BrowserRouter>
-    </ReactQueryConfigProvider>
+            <Route path="/welcome" component={AppWelcome} />
+        </Switch>
+    </BrowserRouter>
 )
 
 ReactDOM.render(routing, document.getElementById('root'))
